+++ src/main/java/com/capstone/ads/model/CustomDesignRequests.java
@@ -25,9 +25,9 @@ public class CustomDesignRequests {
     @GeneratedValue(strategy = GenerationType.UUID)
     String id;
     String requirements;
-    Integer totalPrice;
-    Integer depositAmount;
-    Integer remainingAmount;
+    Long totalPrice;
+    Long depositAmount;
+    Long remainingAmount;
     String finalDesignImage;
     Boolean isNeedSupport;
     Boolean hasOrder;
